+++ examples/sandbox.py
@@ -9,15 +9,19 @@ from vega_datasets import data as vdata
 
 from csvviz.utils.datakit import Datakit
 
+def av(chart):
+     altview.show(chart.interactive())
+
 def main():
-    input_path = vegadata.la_riots.filepath
-    df = pd.read_csv(input_path)
-    dk = Datakit(input_path)
+    la_input_path = vdata.la_riots.filepath
+    ladf = pd.read_csv(la_input_path)
+    ladk = Datakit(la_input_path)
 
+    tk = Datakit('examples/tings.csv')
+    fk = Datakit('examples/fruits.csv')
 
-    fdk = Datakit('examples/fruits.csv')
-
-
+    chart = alt.Chart(tk.df).mark_bar().encode(x='name', y='amount',
+        fill=alt.Color('amount', scale=alt.Scale(range=['green', 'yellow', 'red'] )))
 
     print('Breakpoint time!')
     IBREAKPOINT()